Automatyczny Stock Market 1.0/requirements.txt
# Core runtime
fastapi>=0.110.0,<0.125.0
uvicorn[standard]>=0.24.0
httpx>=0.27.0
PyJWT>=2.8.0
prometheus-client>=0.20.0
Jinja2
python-multipart
authlib
passlib[argon2]
bcrypt
email-validator

# Utilities
python-dotenv>=1.0.1
requests>=2.32.3
typer>=0.12.3
rich>=13.7.1

# AI & Machine Learning
google-generativeai>=0.5.4
scikit-learn>=1.3.0

# Web & API
aiohttp>=3.8.0
httpx>=0.25.0,<0.28.0
websockets

# Data Serialization
orjson>=3.9.0

# Monitoring
structlog>=23.2.0
opentelemetry-sdk>=1.36.0
opentelemetry-api>=1.36.0
opentelemetry-exporter-otlp>=1.36.0
opentelemetry-instrumentation-fastapi>=0.57b0
opentelemetry-instrumentation-sqlalchemy>=0.57b0
opentelemetry-instrumentation-redis>=0.57b0
opentelemetry-instrumentation-logging>=0.57b0
opentelemetry-instrumentation-system-metrics>=0.57b0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
redis
slowapi>=0.1.9

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
isort>=5.12.0
asyncio-throttle>=1.0.0

# Added for web search
beautifulsoup4>=4.11.1
duckduckgo_search>=4.0.0

# Other
ccxt
numpy
pandas
cryptography
ta


